+++ example/template/index.ts
@@ -1,4 +1,4 @@
-import { app, get } from '../../dinatra.ts';
+import { app, get } from '../../mod.ts';
 import { cwd, open } from 'deno';
 
 const currentDir = cwd();
